Add files via upload
Starter backend code for the webpage. Features a grid and a hide class, as well as basic placements for the first pages elements. Now going to work on a better way of doing it than having to write out /span every time to size cells, ideally hoping for a less unsafe structure